[Form] Extracted ReferencingArrayIterator out of VirtualFormAwareIterator
